% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapport.R
\name{all_ggplot_to_docx}
\alias{all_ggplot_to_docx}
\title{all_ggplot_to_docx}
\usage{
all_ggplot_to_docx(laliste = get_plot_to_save(), filename = "rapport.docx",
  open = TRUE, dossier = tempdir())
}
\arguments{
\item{laliste}{liste des graphiques mettre dans le rapport}

\item{filename}{nom du fichier de sortie}

\item{open}{booleen ouverture du fichier apres creation}

\item{dossier}{chemin du dossier de creation}
}
\description{
recupere tous les ggplot in mémory pour en fare un docx et des png
}
