% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapport_rmd.R
\name{all_ggplot_to_rmd_html}
\alias{all_ggplot_to_rmd_html}
\title{all_ggplot_to_Rmd}
\usage{
all_ggplot_to_rmd_html(laliste = get_plot_to_save(),
  filename = "rapport.html", open = TRUE, dossier = tempdir())
}
\arguments{
\item{laliste}{liste des graphiques mettre dans le rapport}

\item{filename}{nom du fichier de sortie}

\item{open}{booleen ouverture du fichier apres creation}

\item{dossier}{chemin du dossier de creation}
}
\description{
recupere tous les ggplot in mémory pour en fare un docx et des png
}
