% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_to_png.R
\name{dygraph_to_png}
\alias{dygraph_to_png}
\title{transforme un objet dygraph en png}
\usage{
dygraph_to_png(objet, filename = "dygraph.png", ...)
}
\arguments{
\item{objet}{un graphique dygraph}

\item{filename}{nom du fichier de sortie}

\item{...}{pas utilise}
}
\description{
transforme un objet dygraph en png
}
