% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot_to_save.R
\name{genere}
\alias{genere}
\title{fabrique le graphique attendu et le met dans .GlobalEnv}
\usage{
genere(todo, ...)
}
\arguments{
\item{todo}{nom du graphique}
}
\description{
fabrique le graphique attendu et le met dans .GlobalEnv
}
