% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapport.R
\name{genere_graph_manquant}
\alias{genere_graph_manquant}
\title{genere_graph_manquant}
\usage{
genere_graph_manquant(a_generer, laliste = laliste)
}
\arguments{
\item{a_generer}{liste des graphiques a generer}

\item{laliste}{liste des graphiques}
}
\description{
genere les graph demandé poru un rapport s'il n'ont pas été fabriqué dans shiny
}
