% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapport.R
\name{graph_a_generer}
\alias{graph_a_generer}
\title{graph_a_generer}
\usage{
graph_a_generer(laliste_avec_dependance)
}
\arguments{
\item{laliste_avec_dependance}{liste des graphiques}
}
\description{
donne la liste des graph a generer ( cad ceux quin'existent pas dans .Globalenv)
}
