% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapport.R
\name{pimp_dependance}
\alias{pimp_dependance}
\title{pimp_dependance}
\usage{
pimp_dependance(laliste)
}
\arguments{
\item{laliste}{liste des graphiques a exporter}
}
\description{
rajoute dans laliste les dependance qui manqueraient peut etre (exemple pour le graph 15 IL FAUT le graph 14)
}
