% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_to_png.R
\name{plotly_to_png}
\alias{plotly_to_png}
\title{transforme un objet plotly en png}
\usage{
plotly_to_png(objet, filename = "plotly.png", ...)
}
\arguments{
\item{objet}{un graphique plotly}

\item{filename}{nom du fichier de sortie}

\item{...}{pas utilise}
}
\description{
transforme un objet plotly en png
}
