% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapport.R
\name{png_to_zip}
\alias{png_to_zip}
\title{png_to_zip}
\usage{
png_to_zip(dossier, nom = "graphiques.zip")
}
\arguments{
\item{dossier}{dossier à compresser}

\item{nom}{nom du fichier zip}
}
\description{
Fabrique un zip depuis le dossier de png
}
