% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapport.R
\name{prepare_pour_rapport}
\alias{prepare_pour_rapport}
\title{prepare_pour_rapport}
\usage{
prepare_pour_rapport(laliste)
}
\arguments{
\item{laliste}{liste des graphiques}
}
\description{
ne renvoie que les truc pour le rapport
}
