% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_to_png.R
\name{rcharts_to_png}
\alias{rcharts_to_png}
\title{transforme un objet rCharts en png}
\usage{
rcharts_to_png(objet, filename = "rchart.png", ...)
}
\arguments{
\item{objet}{un graphique plotly}

\item{filename}{nom du fichier de sortie}

\item{...}{pas utilise}
}
\description{
transforme un objet rCharts en png
}
