% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapport.R
\name{save_to_docx}
\alias{save_to_docx}
\title{save_to_docx}
\usage{
save_to_docx(laliste_rapport, dossier, filename)
}
\arguments{
\item{laliste_rapport}{liste des graphiques}

\item{dossier}{dossier de creation du rapport}

\item{filename}{nom du fichier de sortie}
}
\description{
fabrique le docx  a partir de laliste_rapport dans dossier
}
