% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapport.R
\name{save_to_png}
\alias{save_to_png}
\title{save_to_png}
\usage{
save_to_png(laliste, dossier)
}
\arguments{
\item{laliste}{liste contenant le nom des fichiers a sauvegarder}

\item{dossier}{dossier de sortie}
}
\description{
fabrique les png a partir de laliste dans dossier ( ne gere pas les graph Rbase)
}
