% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapport_rmd.R
\name{save_to_rmd_pdf}
\alias{save_to_rmd_pdf}
\title{save_to_rmds_html}
\usage{
save_to_rmd_pdf(laliste_rapport, dossier, filename,
  output_format = "html_document")
}
\arguments{
\item{laliste_rapport}{liste des graphiques}

\item{dossier}{dossier de creation du rapport}

\item{filename}{nom du fichier de sortie}
}
\description{
fabrique le html  a partir de laliste_rapport dans dossier
}
