% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedometer.R
\name{speedometer}
\alias{speedometer}
\title{speedometer}
\usage{
speedometer(x, titre_jauge = "")
}
\arguments{
\item{x}{indicateur, numeric}

\item{titre_jauge}{une chaine de caracteres}
}
\description{
speedometer
}
\examples{
speedometer(42,titre_jauge ='titre')
}
